Fix tailwind.config.js pathing
With correct pathing tailwind can now generate the
classes used in the html pages to the main.css.

In this case unused classes will not be present in
the stylesheet.

The previously stored main.css is now deleted along
with the tailwind.min.css that was used for the first
setup of the website.

GH-10